# Template abstraction performance