--- a/src/template/html_parser.js
+++ b/src/template/html_parser.js
@@ -5,25 +5,32 @@ var entityMap = require('html-tokenizer/entity-map');
 // @TODO examine other tokenizers or parsers
 var Parser = require('html-tokenizer/parser');
 
-module.exports = function(html, stack) {
-  var _stack = stack || new DefaultStack(true);
+var defaultStack = new DefaultStack(true);
+var _stack;
+var parser = new Parser({
+  entities: entityMap
+});
+parser.on('open',  function(name, attributes) {
+  _stack.openElement(name, attributes);
+});
+parser.on('comment', function(text) {
+  _stack.createComment(text);
+});
+parser.on('text', function(text) {
+  _stack.createObject(text);
+});
+parser.on('close', function() {
+  var el = _stack.peek();
+  _stack.closeElement(el);
+});
 
-  var parser = new Parser({
-    entities: entityMap
-  });
-  parser.on('open',  function(name, attributes) {
-    _stack.openElement(name, attributes);
-  });
-  parser.on('comment', function(text) {
-    _stack.createComment(text);
-  });
-  parser.on('text', function(text) {
-    _stack.createObject(text);
-  });
-  parser.on('close', function() {
-    var el = _stack.peek();
-    _stack.closeElement(el);
-  });
+module.exports = function(html, stack) {
+  if (stack) {
+    _stack = stack;
+  } else {
+    defaultStack.clear();
+    _stack = defaultStack;
+  }
 
   parser.parse(html);