This fixes the majority of the performance issues that were introduced by the template abstraction merge. As identified earlier today.

I created a test template along with a performance timer and came up with the following measurements

Before Template Abstraction (based on commit 2591ee60e614f3059ba346fe66c1ad54c4e2bc18):
```perf_run.rendered: 94.9150ms (949.1502ms / 10)```
Template Abstraction (current master 47362871e7f58a05904df9283870fde06c82ce41)
```perf_run.rendered: 2609.7290ms (26097.2904ms / 10)```

Timers were used to track the performance hit back to the html_parser module, meaning that the issue primarily affected rendering of triple curly data with HTML content as well as dynamic attributes. Parser was being constructed inside to allow various stacks to be passed in, but this was easily adjusted to be constructed as it was previously and use an external variable for the stack. As a result, this PR returns performance to the original state:
```perf_run.rendered: 95.8699ms (958.6990ms / 10)```