@MattDeeg 2015-11-11
It would be good to see if we can work the performance test into the continuous integration suite so that regressions like this are more easily caught